requirements.txt
requests>=2.32.0
beautifulsoup4>=4.14.0
openai>=2.0.0
faiss-cpu>=1.12.0
python-dotenv>=1.0.0
flask>=2.3.0
